Unify credentials and authenticated request midllewares
